+++ packages/core/src/lib/cache/local-cache-provider.ts
@@ -1,4 +1,4 @@
-import type { JsonFeed } from "../parse/parse-xml-feed";
+import type { JsonFeed } from "../json-feed";
 
 import { readdir, readFile, writeFile, mkdir } from "fs/promises";
 import * as uuid from "uuid"; // TODO implement v5 without lib
