+++ packages/core/src/lib/json-feed.d.ts
