+++ packages/core/src/lib/merge/merge-json-feeds.ts
@@ -1,4 +1,4 @@
-import type { JsonFeed } from "../parse/parse-xml-feed";
+import type { JsonFeed } from "../json-feed";
 
 export function mergeJsonFeeds(newFeeds: JsonFeed[], existingFeeds: JsonFeed[]): JsonFeed[] {
   return [...existingFeeds, ...newFeeds]; // TODO implement actual merge with TDD
